// vhome_mis/src/app/core/api-client-v2/model/buildingResponseBuildingDetails.ts
/**
 * APIFlask
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { BuildingResponseFloorDetails } from './buildingResponseFloorDetails';


export interface BuildingResponseBuildingDetails { 
    ID: number;
    address: string;
    building_floors: Array<BuildingResponseFloorDetails>;
    name: string;
}