// vhome_mis/src/app/core/api-client-v2/model/buildingResponseFloorDetails.ts
/**
 * APIFlask
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { BuildingResponseRoomDetails } from './buildingResponseRoomDetails';


export interface BuildingResponseFloorDetails { 
    ID: number;
    floor_name: string;
    living_space: Array<BuildingResponseRoomDetails>;
}