// vhome_mis/src/app/core/api-client-v2/model/buildingResponseRoomDetails.ts
/**
 * APIFlask
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { BuildingResponseServiceUserBasicData } from './buildingResponseServiceUserBasicData';


export interface BuildingResponseRoomDetails { 
    ID: number;
    name: string;
    service_user: Array<BuildingResponseServiceUserBasicData>;
}